Reorganize: one module per reporter under cli/reporters
`check.ts` is now a thin dispatcher. The two reporter bodies it used to inline
move into dedicated modules:

- cli/reporters/pretty.ts        — prettyReporter (human output)
- cli/reporters/json.ts          — jsonReporter (json + ndjson)
- cli/reporters/jsonReportRow.ts — toJsonReportRow + JsonReportRow (renamed
                                   from jsonReport.ts), the shared row shape

No behaviour change.
