+++ src/cli/reporters/jsonReportRow.ts
